Convert the project to multigroup
This will support using both workloads.cloudfoundfoundry.org and
networking.cloudfoundry.org
