+++ apis/workloads/v1alpha1/cfapp_webhook_integration_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"code.cloudfoundry.org/cf-k8s-controllers/api/v1alpha1"
+	"code.cloudfoundry.org/cf-k8s-controllers/apis/workloads/v1alpha1"
 	. "github.com/onsi/gomega"
 	"github.com/sclevine/spec"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
