+++ webhooks/workloads/cfapp_validation.go
@@ -1,7 +1,7 @@
-package webhooks
+package workloads
 
 import (
-	"code.cloudfoundry.org/cf-k8s-controllers/api/v1alpha1"
+	"code.cloudfoundry.org/cf-k8s-controllers/apis/workloads/v1alpha1"
 	"context"
 	v1 "k8s.io/api/admission/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
